# packages/api-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "drymem",
    "description": "Shared long-term memory for AI coding agents.\n\nEvery memory belongs to a project, identified by its normalised git remote, and\nto the person who saved it. Memories are private until explicitly promoted.",
    "version": "2.1.0"
  },
  "paths": {
    "/v1/memories": {
      "post": {
        "tags": [
          "memories"
        ],
        "summary": "Save Memory",
        "description": "Save a session summary. Returns 422 if it contains a private key.",
        "operationId": "save_memory_v1_memories_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveMemoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveMemoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/memories/{topic_key}": {
      "patch": {
        "tags": [
          "memories"
        ],
        "summary": "Update Memory",
        "description": "Append to a topic, or replace it.\n\nAppending is the default because Graphiti reconciles contradictions itself:\nthe older fact is marked superseded rather than lost, which keeps the history\nof a decision readable.",
        "operationId": "update_memory_v1_memories__topic_key__patch",
        "parameters": [
          {
            "name": "topic_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Topic Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveMemoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/memories/search": {
      "get": {
        "tags": [
          "memories"
        ],
        "summary": "Search Memories",
        "operationId": "search_memories_v1_memories_search_get",
        "parameters": [
          {
            "name": "project_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Key"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/memories/context": {
      "get": {
        "tags": [
          "memories"
        ],
        "summary": "Memory Context",
        "operationId": "memory_context_v1_memories_context_get",
        "parameters": [
          {
            "name": "project_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Key"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/memories/topics": {
      "get": {
        "tags": [
          "memories"
        ],
        "summary": "List Topics",
        "description": "Every topic key already stored. Used by `drymem import` to stay idempotent.",
        "operationId": "list_topics_v1_memories_topics_get",
        "parameters": [
          {
            "name": "project_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/memories/{episode_uuid}": {
      "delete": {
        "tags": [
          "memories"
        ],
        "summary": "Delete Memory",
        "description": "Delete a memory. A uuid outside the caller's org is simply not found.",
        "operationId": "delete_memory_v1_memories__episode_uuid__delete",
        "parameters": [
          {
            "name": "episode_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Episode Uuid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/memories/{episode_uuid}/feedback": {
      "post": {
        "tags": [
          "memories"
        ],
        "summary": "Rate Memory",
        "description": "Record whether a retrieved memory was useful.\n\nThis is the pilot's precision metric: PLAN.md's 90% bar is measured from\nthese rows, so the query that surfaced the memory is stored alongside the\nthumb \u2014 a rating with no query cannot be learned from.",
        "operationId": "rate_memory_v1_memories__episode_uuid__feedback_post",
        "parameters": [
          {
            "name": "episode_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Episode Uuid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/memories/{episode_uuid}/promote": {
      "post": {
        "tags": [
          "memories"
        ],
        "summary": "Promote Memory",
        "description": "Share a memory with the project's members.\n\nOnly your own memory, and only into a project you belong to. Promoting an\nalready-promoted memory succeeds without duplicating it, so a retry is safe.",
        "operationId": "promote_memory_v1_memories__episode_uuid__promote_post",
        "parameters": [
          {
            "name": "episode_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Episode Uuid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_key}/members": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Members",
        "operationId": "list_members_v1_projects__project_key__members_get",
        "parameters": [
          {
            "name": "project_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Add Member",
        "description": "Add someone to a project. They see its team memories, never its private ones.",
        "operationId": "add_member_v1_projects__project_key__members_post",
        "parameters": [
          {
            "name": "project_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SaveMemoryRequest": {
        "properties": {
          "project_key": {
            "type": "string",
            "title": "Project Key",
            "description": "Normalised git remote, e.g. github.com/acme/payments"
          },
          "summary": {
            "type": "string",
            "minLength": 1,
            "title": "Summary",
            "description": "Markdown body written by the agent"
          },
          "topic_key": {
            "type": "string",
            "title": "Topic Key",
            "description": "Stable key for cross-session linking",
            "default": ""
          },
          "tool": {
            "type": "string",
            "title": "Tool",
            "description": "Which agent produced this",
            "default": "claude-code"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "decision \u00b7 architecture \u00b7 bugfix \u00b7 discovery \u00b7 convention \u00b7 note. An unknown value becomes 'note' rather than failing the save.",
            "default": "note"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "The agent run this came out of",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "project_key",
          "summary"
        ],
        "title": "SaveMemoryRequest"
      },
...
      "SearchResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "project_key": {
            "type": "string",
            "title": "Project Key"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/FactOut"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "query",
          "project_key",
          "results"
        ],
        "title": "SearchResponse"
      },
...
    }
  }
}